refactor(assistance): Correct viewset dir name to viwsets
ref: #248 #352
